Factor node stuff out into new file
