+++ io_scene_gltf/__init__.py
@@ -3,15 +3,13 @@ import os
 import struct
 
 import bpy
-from mathutils import Matrix
-from mathutils import Quaternion
-from mathutils import Vector
 from bpy.props import StringProperty
 from bpy_extras.io_utils import ImportHelper
 
 from io_scene_gltf import buffer
 from io_scene_gltf import material
 from io_scene_gltf import mesh
+from io_scene_gltf import node
 
 bl_info = {
     "name": "glTF 2.0 Importer",
@@ -24,15 +22,6 @@ bl_info = {
     "category": "Import-Export"
 }
 
-def convert_matrix(m):
-    """Convert glTF matrix to Blender matrix"""
-    result = Matrix([m[0:4], m[4:8], m[8:12], m[12:16]])
-    result.transpose() # column-major to row-major
-    return result
-
-def convert_quaternion(q):
-    """Convert glTF quaternion to Blender quaternion"""
-    return Quaternion([q[3], q[0], q[1], q[2]]) # xyzw -> wxyz
 
 class ImportGLTF(bpy.types.Operator, ImportHelper):
     bl_idname = "import_scene.gltf"
@@ -65,27 +54,6 @@ class ImportGLTF(bpy.types.Operator, ImportHelper):
             self.default_material = material.create_default_material()
         return self.default_material
 
-    def set_transform(self, obj, node):
-        if 'matrix' in node:
-            obj.matrix_local = convert_matrix(node['matrix'])
-        else:
-            mat = Matrix()
-            if 'scale' in node:
-                s = node['scale']
-                mat = Matrix([
-                    [s[0], 0, 0, 0],
-                    [0, s[1], 0, 0],
-                    [0, 0, s[2], 0],
-                    [0, 0, 0, 1]
-                ])
-            if 'rotation' in node:
-                q = convert_quaternion(node['rotation'])
-                mat = q.to_matrix().to_4x4() * mat
-            if 'translation' in node:
-                t = Vector(node['translation'])
-                mat = Matrix.Translation(t) * mat
-            obj.matrix_local = mat
-
     def get_mesh(self, idx):
         if idx not in self.meshes:
             self.meshes[idx] = mesh.create_mesh(self, idx)
@@ -99,54 +67,10 @@ class ImportGLTF(bpy.types.Operator, ImportHelper):
             self.cameras[idx] = bpy.data.cameras.new(name)
         return self.cameras[idx]
 
-    def create_object(self, idx, parent, scene):
-        node = self.root['nodes'][idx]
-        name = node.get('name', 'nodes[%d]' % idx)
-        ob = bpy.data.objects.new(name, None)
-
-        if 'mesh' in node:
-            mesh_ob = bpy.data.objects.new(
-                name + '.mesh',
-                self.get_mesh(node['mesh'])
-            )
-            mesh_ob.parent = ob
-            scene.objects.link(mesh_ob)
-        if 'camera' in node:
-            camera_ob = bpy.data.objects.new(
-                name + '.camera',
-                self.get_camera(node['camera'])
-            )
-            camera_ob.parent = ob
-            scene.objects.link(camera_ob)
-        #TODO handle skin
-
-        self.set_transform(ob, node)
-
-        ob.parent = parent
-        bpy.context.scene.objects.link(ob)
-        scene.update()
-
-        if 'children' in node:
-            children = node['children']
-            for idx in children:
-                self.create_object(idx, ob, scene)
-
-    def create_scene(self, idx):
-        scene = self.root['scenes'][idx]
-        name = scene.get('name', 'scene[%d]' % idx)
-
-        bpy.ops.scene.new(type = 'NEW')
-        scn = bpy.context.scene
-        scn.name = name
-        scn.render.engine = 'CYCLES'
-        # scn.world.use_nodes = True
-
-        for root_idx in scene.get('nodes', []):
-            self.create_object(root_idx, None, scn)
-
-        scn.update()
-
-        self.scenes[idx] = scn
+    def generate_scenes(self):
+        if 'scenes' in self.root:
+            for scene_idx in range(0, len(self.root['scenes'])):
+                node.create_scene(self, scene_idx)
 
     def check_version(self):
         def string_to_version(s):
@@ -206,9 +130,7 @@ class ImportGLTF(bpy.types.Operator, ImportHelper):
 
         self.check_version()
 
-        if 'scenes' in self.root:
-            for scene_idx in range(0, len(self.root['scenes'])):
-                self.create_scene(scene_idx)
+        self.generate_scenes()
 
         if 'scene' in self.root:
             bpy.context.screen.scene = self.scenes[self.root['scene']]
